extract conflicting test to another file
